build: add build.sh convenience script for linux
This makes it easier to do rebuilds using docker, because it doesn't
actually modify the source tree (so that Git actually understands what
you are doing with the chmodding), as all directories touched by docker
are temporary, except for `result`, which is the "output". This just
contains ISLE.EXE, LEGO1.DLL and LEGO1.lib.

Optionally Nix can be used to speed up the "copy the source to somewhere
temporary" step.

Note that this doesn't work for MacOS, because this script uses
`--tmpfs`, which is Linux-only.